validate openai keys on embeddings creation
